Convert point ids from Int to Long
This allows larger datasets and plays more nicely with `RDD.zipWithIndex`.

+++ src/main/scala/com/github/karlhigley/spark/neighbors/lsh/MinhashFunction.scala
@@ -35,7 +35,7 @@ private[neighbors] class MinhashFunction(
   /**
    * Build a hash table entry for the supplied vector
    */
-  def hashTableEntry(id: Int, table: Int, v: SparseVector): IntHashTableEntry = {
+  def hashTableEntry(id: Long, table: Int, v: SparseVector): IntHashTableEntry = {
     IntHashTableEntry(id, table, signature(v), v)
   }
 }
